refactor: renamed directories to improve readability
Directory 'home' was renamed to 'common', implying common dotfiles
regardless of distro or environment (that's the idea but so far only
Debian and WSL2 Debian have been validated). Also, specific
configuration was moved to the respective directory, and WSL2 Docker
config was refactored to 'dot-profile' rather than 'dot-bashrc' since it
is independent of Bash.